Rearrange folders
Fix all linter problems
Add test coverage
Remove set folder - see go-set repo
