+++ .github/workflows/ci.yml
@@ -7,12 +7,9 @@ on:
     branches:
       - main
 
-permissions:
-  contents: write # to fetch code (actions/checkout)
-
 jobs:
   vulncheck:
-    name: Analysis
+    name: CI
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
