+++ apis/widgets/widgets.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 package widgetsapi;
-option go_package="github.com/eccles/hestia/pkg/api/widgets;widgetsapi";
+option go_package="github.com/eccles/hestia/api/widgets;widgetsapi";
 
 import "google/protobuf/empty.proto";
 import "google/protobuf/field_mask.proto";
