+++ cmd/hestia/cmd/root.go
@@ -1,17 +1,28 @@
-// Entrypoint for hestia command (no subcommands)
+// .*@mycompany\.com MY COMPANY 2025
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package cmd
 
 import (
 	"fmt"
-	"math/rand"
 	"os"
-	"time"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/eccles/hestia/pkg/logger"
+	"github.com/eccles/hestia/logger"
 )
 
 const (
@@ -42,6 +53,7 @@ func parseRootCmd() {
 func Execute() {
 	rootNew()
 	widgetsNew()
+
 	if err := rootCmd.Execute(); err != nil {
 		os.Exit(1)
 	}
@@ -66,8 +78,6 @@ func rootNew() {
 	)
 
 	_ = viper.BindPFlag(logLevelLabel, rootCmd.PersistentFlags().Lookup(logLevelLabel))
-
-	rand.New(rand.NewSource(time.Now().UnixNano()))
 }
 
 // initConfig reads in config file and ENV variables if set.
@@ -89,6 +99,7 @@ func initConfig() {
 	if err := viper.ReadInConfig(); err == nil {
 		fmt.Println("Using config file:", viper.ConfigFileUsed())
 	}
+
 	if err := viper.WriteConfigAs("." + rootCmd.Use + ".yaml"); err != nil {
 		fmt.Println("Cannot write config file:", err)
 	}
