--- pkg/apis/widgets/generate.go
@@ -1,4 +1,0 @@
-package widgetsapi
-
-// Generate Go code from protobuf definition file.
-//go:generate ../../../scripts/protoc.sh pkg/apis/widgets
