--- pkg/startup/logger.go
@@ -1,7 +1,0 @@
-package startup
-
-import (
-	"github.com/eccles/hestia/pkg/logger"
-)
-
-type Logger = logger.Logger
