--- pkg/widgets/service/logger.go
@@ -1,7 +1,0 @@
-package widgetsservice
-
-import (
-	"github.com/eccles/hestia/pkg/logger"
-)
-
-type Logger = logger.Logger
