+++ startup/terminate.go
@@ -13,13 +13,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package startup
 
 import (
-	"github.com/eccles/hestia/startup"
-	widgetsservice "github.com/eccles/hestia/widgets/service"
+	"os"
+	"os/signal"
+	"syscall"
 )
 
-func main() {
-	startup.Run("widgets", widgetsservice.Run)
+func WaitForTermination() {
+	// wait here for termination signal
+	sCh := make(chan os.Signal, 1)
+
+	signal.Notify(sCh, syscall.SIGINT, syscall.SIGTERM)
+	<-sCh
+
+	// k8s is in charge now so undo handling of signals
+	signal.Reset(syscall.SIGINT, syscall.SIGTERM)
 }
